cleanup: remove aws configure step
Creating an AWS profile was a legacy of the initial impl which was based
off of the s3-sync-action which does something similar. If AWS creds
(via access/secret pair or OIDC session token) and region are available
in the environment, then there is no real reason to create a profile to
run commands with, as documented here:

https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-envvars.html

This is simplified even more now that there is a supported action for
configuring AWS creds before running others.